Move prefpane files into "prefpane" subdir
Note that we do not move the project file, only the component files. The
reason is that we want the project file to live at the same level (ie.
share the SRCROOT) with the Fusion plug-in that is going to embed the
preference pane. This will make things like relative paths to the built
products directory more straightforward.

Signed-off-by: Wincent Colaiuta <win@wincent.com>